backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.37.37
botocore==1.37.37
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.12
greenlet==3.2.0
h11==0.14.0
idna==3.10
jmespath==1.0.1
joblib==1.4.2
numpy==2.2.5
pandas==2.2.3
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
s3transfer==0.11.5
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
threadpoolctl==3.6.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
